@@ -11,45 +11,42 @@
     "jest.test": "jest"
   },
   "devDependencies": {
-    "@types/history": "^4.7.9",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-redux": "^7.1.18",
-    "@types/react-router-dom": "^5.3.0",
-    "utility-types": "3.10.0",
-    "babel-jest": "25.1.0",
-    "enzyme": "3.11.0",
-    "enzyme-adapter-react-16": "1.15.5",
-    "eslint": "5.13.0",
+    "@types/history": "5.0.0",
+    "@types/react-dom": "18.0.11",
+    "@types/react-redux": "7.1.25",
+    "@types/react-router-dom": "5.3.3",
+    "babel-jest": "29.4.3",
+    "eslint": "8.34.0",
     "eslint-config-htmlacademy": "0.5.2",
-    "jest": "25.1.0",
-    "react-test-renderer": "16.12.0",
-    "redux-devtools-extension": "2.13.8",
+    "jest": "29.4.3",
+    "jest-environment-jsdom": "^29.4.3",
+    "react-test-renderer": "18.2.0",
+    "redux-devtools-extension": "2.13.9",
     "redux-mock-store": "1.5.4",
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.4.3"
-  },
-  "engines": {
-    "node": "10"
+    "ts-loader": "9.4.2",
+    "typescript": "4.9.5",
+    "utility-types": "3.10.0"
   },
   "dependencies": {
-    "@babel/core": "7.8.7",
-    "@babel/preset-env": "7.8.7",
-    "@babel/preset-react": "7.8.3",
-    "axios": "0.21.4",
-    "axios-mock-adapter": "1.17.0",
-    "babel-loader": "8.1.0",
-    "eslint-plugin-react": "7.18.3",
-    "history": "^4.10.1",
-    "loader-utils": "^2.0.0",
-    "react": "16.12.0",
-    "react-dom": "16.12.0",
-    "react-redux": "7.2.1",
-    "react-router-dom": "5.1.2",
-    "redux": "4.0.5",
-    "redux-saga": "^1.1.3",
-    "reselect": "4.0.0",
-    "webpack": "5.53.0",
-    "webpack-cli": "3.3.12",
-    "webpack-dev-server": "3.10.3"
+    "@babel/core": "7.20.12",
+    "@babel/preset-env": "7.20.2",
+    "@babel/preset-react": "7.18.6",
+    "@testing-library/react": "^13.4.0",
+    "axios": "1.3.3",
+    "axios-mock-adapter": "1.21.2",
+    "babel-loader": "9.1.2",
+    "eslint-plugin-react": "7.32.2",
+    "history": "4.10.1",
+    "loader-utils": "3.2.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-redux": "8.0.5",
+    "react-router-dom": "5.3.0",
+    "redux": "4.2.1",
+    "redux-saga": "1.2.2",
+    "reselect": "4.1.7",
+    "webpack": "5.75.0",
+    "webpack-cli": "5.0.1",
+    "webpack-dev-server": "4.11.1"
   }
 }
