@@ -1,6 +1,6 @@
-import React, {FC} from 'react';
-import {connect} from 'react-redux';
-import {Router as BrowserRouter, Switch, Route, Link, Redirect, RouteChildrenProps} from 'react-router-dom';
+import React, { FC } from 'react';
+import { connect } from 'react-redux';
+import { Router as BrowserRouter, Switch, Route, Link, Redirect, RouteChildrenProps } from 'react-router-dom';
 import Main from '../main/main';
 import Singin from '../singin/singin';
 import Addreview from '../addreview/addreview';
@@ -12,8 +12,8 @@ import Player from '../player/player';
 import withPlayer from '../../hocs/withPlayer/withPlayer';
 import withTags from '../../hocs/with-tags/with-tags';
 import Loading from '../loading/loading';
-import {getIsLoading} from '../../store/reducers/films/films-selectors';
-import {RootState} from '../../store/reducers/root-reducer';
+import { getIsLoading } from '../../store/reducers/films/films-selectors';
+import { RootState } from '../../store/reducers/root-reducer';
 
 const VideoWrappedPlayer = withPlayer(Player);
 const MainWithTags = withTags(Main);
@@ -23,7 +23,7 @@ type Props = {
 };
 
 const App: FC<Props> = (props) => {
-  const {isLoading} = props;
+  const { isLoading } = props;
   return (
     <BrowserRouter history={history}>
       <Switch>
@@ -39,13 +39,13 @@ const App: FC<Props> = (props) => {
         </Route>
         <PrivateRouter
           path="/films/:id/review"
-          render={(routerProps: RouteChildrenProps<{id?: string}>) => {
+          render={(routerProps: RouteChildrenProps<{ id?: string }>) => {
             const selectedID = routerProps?.match?.params?.id;
             if (!selectedID) {
               return <Redirect to="/" />;
             }
             return isLoading ? (
-              <div style={{background: `black`, height: `100vh`}}>
+              <div style={{ background: `black`, height: `100vh` }}>
                 <Loading />
               </div>
             ) : (
@@ -56,13 +56,13 @@ const App: FC<Props> = (props) => {
         <Route
           path="/films/:id"
           exact
-          render={(routerProps: RouteChildrenProps<{id?: string}>) => {
+          render={(routerProps: RouteChildrenProps<{ id?: string }>) => {
             const selectedID = routerProps?.match?.params?.id;
             if (!selectedID) {
               return <Redirect to="/" />;
             }
             return isLoading ? (
-              <div style={{background: `black`, height: `100vh`}}>
+              <div style={{ background: `black`, height: `100vh` }}>
                 <Loading />
               </div>
             ) : (
@@ -78,14 +78,14 @@ const App: FC<Props> = (props) => {
         />
         <Route
           path="/player/:id"
-          render={(routerProps: RouteChildrenProps<{id?: string}>) => {
+          render={(routerProps: RouteChildrenProps<{ id?: string }>) => {
             const selectedID = routerProps?.match?.params?.id;
             console.log(selectedID);
             if (!selectedID) {
               return <Redirect to="/" />;
             }
             return isLoading ? (
-              <div style={{background: 'black', height: '100vh'}}>
+              <div style={{ background: 'black', height: '100vh' }}>
                 <Loading />
               </div>
             ) : (
