@@ -1,14 +1,16 @@
 const path = require('path');
 
 module.exports = {
+  mode: 'development',
   entry: './src/index.tsx',
   output: {
     filename: 'bundle.js',
     path: path.resolve(__dirname, 'public'),
   },
   devServer: {
-    contentBase: path.resolve(__dirname, 'public'),
+    static: path.resolve(__dirname, 'public'),
     historyApiFallback: true,
+    hot: true,
     open: true,
     port: 1337,
   },
